Commit for implementation of multiple histograms in a single run.
This commit implements a mechanism whereby you may have any number
of histograms filled in a single run. This is done by having an STL
vector of pointers to histograms, held by the sensitive detector.
On each step, the sensitive detector loops though the vector and
calls fill on it, with the step information.

In order to keep everything sensible and allow dose and LET histograms
to be done simultaneously, the decision about histogram type has
been delegated into the histogram class (where it probably should
be).

These changes are likely to result in a performance hit, depending
on the number of histograms used. It is also very important that
histograms are given a sensible name now!

Also included are an example file for multiple histograms, and a
2d histogram plotting script, which might be handy soon.